Add react-datepicker (#7)
This PR adds react-datepicker to improve date selection and to fix the timezone issue with the default date input field. I consider it "good enough for now" but not perfect - a custom date picker (maybe based on https://material-ui-pickers.dev/) would let us build something even nicer.

Styling works and is based on the material-ui theme. One quirk is that the picker pops up either below the entry field (for the start time) or above it (for the end time) because there isn't enough room below it. I imagine there will be more components (like a description) below the date fields soon enough and the problem will go away then.